+++ research/enc_brute_force.h
@@ -7,7 +7,7 @@ namespace research {
 
 // MSTから総当たりで、圧縮率が良くなるケースだけひとつの画像にまとめる
 std::vector<EncodedCombinedImage> EncodeWithBruteForce(
-    ImagesProvider &images, std::shared_ptr<const ImageTree<size_t>> root,
+    ImagesProvider &images, std::shared_ptr<const ImageTree<double>> root,
     const jxl::ModularOptions &options, const EncodingOptions &encoding_options,
     ProgressReporter *progress);
